new package, move things around, allow the polyfill to accept a target container to watch
