+++ packages/dev/docs/package.json
@@ -15,12 +15,12 @@
     "@react-spectrum/picker": "^3.0.0",
     "@react-spectrum/text": "^3.0.0",
     "@react-spectrum/view": "^3.0.0",
+    "@react-aria/aria-modal-polyfill": "^3.0.0",
     "@react-aria/interactions": "^3.0.0",
     "@react-aria/visually-hidden": "^3.0.0",
     "@react-spectrum/theme-default": "^3.0.0",
     "@spectrum-icons/ui": "^3.0.0",
     "@spectrum-icons/workflow": "^3.0.0",
-    "aria-hidden": "^1.1.1",
     "highlight.js": "9.18.1",
     "classnames": "^2.2.6",
     "@mdx-js/react": "^1.5.5",
